Separate concerns in cg-producer
Move call graph geenration utilities to a separate file
